Treat `graph_inputs`, `graph_outputs`, and `graph_ctx` as normal nodes to simplify node handling (#66)
Previously, `graph_inputs`, `graph_outputs` and `graph_ctx` were treated as `meta-nodes`. The design is based on the special nature of the meta nodes: their inputs and outputs are the same. However, this make the code base very complex, because we need to manage two node collections:
- normal node collection
- meta node collection

as well as the `meta links` between these two collections. As discussed in https://github.com/aiidateam/aiida-workgraph/issues/574

In this PR, we do not distinguish the meta node and the normal node, and treat `graph_inptus`, `graph_outpus` and `graph_ctx` as normal nodes. To avoid name conflicts, the task names graph_inputs, graph_outputs, and graph_ctx are reserved from the beginning.

Those meta tasks are only added when the user starts using them.