`to_dict()` separates runtime data and metadata (#88)
- Previously, we always exported the inputs and outputs sockets (all the metadata and the values).
In this PR, we only export the input values (line 190). One can set `include_sockets=True` to export the whole sockets metadata, e.g., when visualizing the inputs/outputs in the widget. 
- Because the user can attach `error_handlers` for a task, we consider it as `runtime` data.
Additionally, rename the `executor` to `callable`, and add additional fields for the error handler, which does not belong to this PR, but I don't have time to split the PR.