Rename `node` to `task` (#129)
The term `Node` is ambiguous (can be a task or an ORM node in the database). Renaming it to `Task` clarifies its role within this package, i.e., each node performs a task.
This commit renames
- `Node` -> `Task`
- `NodeGraph` -> `Graph`

I still keep the term `node-graph`, because it is a very standard and well-understood term in computer science, clearly indicating that this package handles graph structures.